Removed errors

+++ 3DProj/3DProj/Core/Renderers/DeferredRenderer.cpp
@@ -37,7 +37,7 @@ DeferredRenderer::DeferredRenderer(Display* display)
 	this->combineBuffer->unbindTexture();
 
 	this->shadowResScale = 4.0f;
-	this->shadowBuffer = new FrameBuffer(display->getWidth()*this->shadowResScale, display->getHeight()*this->shadowResScale);
+	this->shadowBuffer = new FrameBuffer((unsigned int)(display->getWidth()*this->shadowResScale), (unsigned int)(display->getHeight()*this->shadowResScale));
 	this->shadowBuffer->createTextures(std::vector<std::pair<FrameBuffer::FBO_ATTATCHMENT_TYPE, GLuint>>{
 		{ FrameBuffer::FBO_DEPTH_ATTACHMENT, GL_RGBA16F } // Depth from shadow camera.
 	});
@@ -125,7 +125,7 @@ void DeferredRenderer::resize(Display * display)
 	//this->glowBuffer->resize(display->getWidth(), display->getHeight());
 	this->glowFilter->resize(display->getWidth(), display->getHeight());
 	this->blurFilter->resize(display->getWidth(), display->getHeight());
-	this->shadowBuffer->resize(display->getWidth()*this->shadowResScale, display->getHeight()*this->shadowResScale);
+	this->shadowBuffer->resize((unsigned int)(display->getWidth()*this->shadowResScale), (unsigned int)(display->getHeight()*this->shadowResScale));
 }
 
 const FrameBuffer * DeferredRenderer::getGBuffer() const
